+++ src/types.ts
@@ -910,6 +910,7 @@ export type Matcher = 'strict' | 'fuzzy' | 'regex'
 
 export interface ListOptions {
   position: string
+  reverse: boolean
   input: string
   ignorecase: boolean
   interactive: boolean